engine/strategylab/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
httpx==0.26.0
pydantic==2.5.3
python-dateutil==2.8.2
pandas==2.2.0
numpy==1.26.3
redis==5.0.1
python-dotenv==1.0.0
